requirements.txt
# Core dependencies
numpy>=1.24.0

# LLM API clients
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0


